Switched nghbr array from view over all m,n,vals, to seperate DualViews over m, stored in a struct which is created as a fixed array at compile time of size 26 (for uniform grid).  This deisng is the same that will be used for BoundaryBuffers, and fixes problem that different buffers have different sizes, so cannot all be stored in a view over (m,n,vals).

+++ src/hydro/bvals/apply_physical_bcs.cpp
@@ -29,10 +29,10 @@ TaskStatus Hydro::HydroApplyPhysicalBCs(Driver* pdrive, int stage)
   for (int m=0; m<nmb; ++m) {
     // apply physical boundaries to inner_x1
     switch (pmy_pack->pmb->mb_bcs(m,BoundaryFace::inner_x1)) {
-      case static_cast<int>(BoundaryFlag::reflect):
+      case BoundaryFlag::reflect:
         ReflectInnerX1(m);
         break;
-      case static_cast<int>(BoundaryFlag::outflow):
+      case BoundaryFlag::outflow:
         OutflowInnerX1(m);
         break;
       default:
@@ -41,10 +41,10 @@ TaskStatus Hydro::HydroApplyPhysicalBCs(Driver* pdrive, int stage)
 
     // apply physical bounaries to outer_x1
     switch (pmy_pack->pmb->mb_bcs(m,BoundaryFace::outer_x1)) {
-      case static_cast<int>(BoundaryFlag::reflect):
+      case BoundaryFlag::reflect:
         ReflectOuterX1(m);
         break;
-      case static_cast<int>(BoundaryFlag::outflow):
+      case BoundaryFlag::outflow:
         OutflowOuterX1(m);
         break;
       default:
@@ -56,10 +56,10 @@ TaskStatus Hydro::HydroApplyPhysicalBCs(Driver* pdrive, int stage)
   for (int m=0; m<nmb; ++m) {
     // apply physical bounaries to inner_x2
     switch (pmy_pack->pmb->mb_bcs(m,BoundaryFace::inner_x2)) {
-      case static_cast<int>(BoundaryFlag::reflect):
+      case BoundaryFlag::reflect:
         ReflectInnerX2(m);
         break;
-      case static_cast<int>(BoundaryFlag::outflow):
+      case BoundaryFlag::outflow:
         OutflowInnerX2(m);
         break;
       default:
@@ -68,10 +68,10 @@ TaskStatus Hydro::HydroApplyPhysicalBCs(Driver* pdrive, int stage)
 
     // apply physical bounaries to outer_x1
     switch (pmy_pack->pmb->mb_bcs(m,BoundaryFace::outer_x2)) {
-      case static_cast<int>(BoundaryFlag::reflect):
+      case BoundaryFlag::reflect:
         ReflectOuterX2(m);
         break;
-      case static_cast<int>(BoundaryFlag::outflow):
+      case BoundaryFlag::outflow:
         OutflowOuterX2(m);
         break;
       default:
@@ -83,10 +83,10 @@ TaskStatus Hydro::HydroApplyPhysicalBCs(Driver* pdrive, int stage)
   for (int m=0; m<nmb; ++m) {
     // apply physical bounaries to inner_x3
     switch (pmy_pack->pmb->mb_bcs(m,BoundaryFace::inner_x3)) {
-      case static_cast<int>(BoundaryFlag::reflect):
+      case BoundaryFlag::reflect:
         ReflectInnerX3(m);
         break;
-      case static_cast<int>(BoundaryFlag::outflow):
+      case BoundaryFlag::outflow:
         OutflowInnerX3(m);
         break;
       default:
@@ -95,10 +95,10 @@ TaskStatus Hydro::HydroApplyPhysicalBCs(Driver* pdrive, int stage)
 
     // apply physical bounaries to outer_x3
     switch (pmy_pack->pmb->mb_bcs(m,BoundaryFace::outer_x3)) {
-      case static_cast<int>(BoundaryFlag::reflect):
+      case BoundaryFlag::reflect:
         ReflectOuterX3(m);
         break;
-      case static_cast<int>(BoundaryFlag::outflow):
+      case BoundaryFlag::outflow:
         OutflowOuterX3(m);
         break;
       default:
